refactor(client): extract column value parsing into dedicated module
Move ~1,650 lines of column/row parsing logic from the monolithic
client.rs (4,589 lines) into a new column_parser.rs module. This
includes parse_column_value, PLP decoders, SQL_VARIANT parsing,
money/date/time helpers, and row conversion functions.

All extracted functions are pure (no self access) and become pub(crate)
free functions. Zero public API changes. 12 unit tests moved with the
code.

client.rs: 4,589 → 2,941 lines
column_parser.rs: 1,664 lines (new)
